requirements.txt
flask==3.0.2
python-dotenv==1.0.1
google-generativeai==0.3.2
twilio==8.12.0
gunicorn==21.2.0
psycopg2-binary==2.9.9
SQLAlchemy>=1.4,<2.0
APScheduler==3.10.4
pytz==2023.3 